% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_dist}
\alias{mean_dist}
\title{Mean euclidean distance in parallel}
\usage{
mean_dist(y)
}
\arguments{
\item{y}{NumericMatrix p x n data matrix}
}
\value{
res list storing Q the n x n unormalized euclidian distance matrix 
        and M the n x n matrix storing the number of non-missing features
}
\description{
Computes the mean euclidian distance matrix
             for all pairs of samples accross all the features.
             This only includes non-missing feature for both pairs.
}
