% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_dist_pair}
\alias{mean_dist_pair}
\title{Mean euclidean distance for a single pair of observations}
\usage{
mean_dist_pair(y_i, y_j)
}
\arguments{
\item{y_i}{NumericVector p data vector for sample i}

\item{y_j}{NumericVector p data vector for sample j}
}
\value{
d_ij euclidian distance normalized by number of non-missing
        features
}
\description{
computes the mean euclidian distance
             for a single pair of samples all features
}
