% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_geno}
\alias{read_geno}
\title{Read genotype data}
\usage{
read_geno(geno_path, n)
}
\arguments{
\item{geno_path}{string of path to eigenstrat geno file}

\item{n}{int number of columns or samples in the geno file}
}
\value{
Y matrix of genotypes
}
