% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{sample_hets}
\alias{sample_hets}
\title{Sample heterozygotes}
\usage{
sample_hets(Y, idx)
}
\arguments{
\item{Y}{matrix n x p of genotypes}

\item{idx}{row indicies of people you want to het sample}
}
\value{
Y_sampled matrix of het sampled genotypes
}
