% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{Times two}
\usage{
timesTwo(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\description{
a description about the function
             note this is a dummy function for help
}
