% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mean_dist}
\alias{mean_dist}
\title{Mean euclidean distance}
\usage{
mean_dist(y)
}
\arguments{
\item{y}{NumericMatrix n x p data matrix}
}
\value{
d n x n euclidian distance matrix
}
\description{
Computes the mean euclidian distance matrix
             for all pairs of samples accross all the features.
             This only includes non-missing feature for both pairs.
}
